backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic
pydantic[email]
pydantic-settings
google-genai
qdrant-client
langgraph
langchain-core
langchain-google-genai
passlib[bcrypt]
bcrypt
python-jose[cryptography]
python-multipart

